back/requirements.txt
Django==3.2
djangorestframework
curl
httpie
dj-database-url==0.5.0
django-filter==2.3.0
gunicorn==20.0.4
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
django-cors-headers
xhtml2pdf
django-userforeignkey
djangorestframework-simplejwt
pylint
django-admin-interface
django-rest-swagger
coreapi
django-coreapi
freeze
reportlab
fpdf2
django-nested-admin




